+++ CHANGES.md
@@ -11,6 +11,7 @@
 - New config setting [show_client_info](https://conversejs.org/docs/html/configuration.html#show-client-info)
 - #1149: With `xhr_user_search_url`, contact requests are not being sent out
 - #1213: Switch roster filter input and icons
+- #1327: fix False mentions positives in URLs and Email addresses
 - #1373: Re-add support for the [muc_domain](https://conversejs.org/docs/html/configuration.html#muc-domain) setting
 - #1400: When a chat message is just an emoji, enlarge the emoji
 - #1437: List of groupchats in modal doesn't scroll